create and wire up WeatherWeek fragment

+++ app/src/main/java/com/toddburgessmedia/gdgweatherapp/data/week/Main.kt
@@ -11,7 +11,7 @@ data class Main(
     val pressure: Int,
     val sea_level: Int,
     val temp: Double,
-    val temp_kf: Int,
+    val temp_kf: Double,
     val temp_max: Double,
     val temp_min: Double
 ) : Parcelable
